pyproject.toml
[project]
name = "aida"
version = "0.1.0"
description = "Asistente Inteligente de Despacho Aduanero"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-formrecognizer>=3.3.3",
    "azure-cognitiveservices-speech>=1.49.1",
    "duckdb>=1.5.2",
    "fastapi>=0.136.1",
    "google-cloud-dialogflow>=2.47.0",
    "pyngrok>=8.1.2",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "uvicorn>=0.46.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
